refactor: separate tool types and labels
